+++ src/cctx.h
@@ -1,0 +1,3 @@
+
+ZSTD_CCtx *init_cctx(int level, int num_threads);
+ZSTD_CCtx * external_ptr_to_zstd_context(SEXP cctx_);