+++ src/init.c
@@ -2,12 +2,14 @@
 #include <R.h>
 #include <Rinternals.h>
 
-extern SEXP zstd_serialize_(SEXP robj_, SEXP level_, SEXP num_threads_);
+extern SEXP init_cctx_(SEXP level_, SEXP num_threads_);
+
+extern SEXP zstd_serialize_(SEXP robj_, SEXP level_, SEXP num_threads_, SEXP cctx_);
 extern SEXP zstd_unserialize_(SEXP raw_vec_);
 
-extern SEXP zstd_serialize_stream_(SEXP robj_, SEXP level_, SEXP num_threads_);
+extern SEXP zstd_serialize_stream_(SEXP robj_, SEXP level_, SEXP num_threads_, SEXP cctx_);
 
-extern SEXP zstd_compress_(SEXP raw_vec_, SEXP level_, SEXP num_threads_);
+extern SEXP zstd_compress_(SEXP raw_vec_, SEXP level_, SEXP num_threads_, SEXP cctx_);
 extern SEXP zstd_decompress_(SEXP raw_vec_);
 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -15,10 +17,11 @@ extern SEXP zstd_decompress_(SEXP raw_vec_);
 // .Call   R_CallMethodDef
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 static const R_CallMethodDef CEntries[] = {
-  {"zstd_serialize_"       , (DL_FUNC) &zstd_serialize_       , 3},
-  {"zstd_serialize_stream_", (DL_FUNC) &zstd_serialize_stream_, 3},
+  {"init_cctx_"            , (DL_FUNC) &init_cctx_            , 2},
+  {"zstd_serialize_"       , (DL_FUNC) &zstd_serialize_       , 4},
+  {"zstd_serialize_stream_", (DL_FUNC) &zstd_serialize_stream_, 4},
   {"zstd_unserialize_"     , (DL_FUNC) &zstd_unserialize_     , 1},
-  {"zstd_compress_"        , (DL_FUNC) &zstd_compress_        , 3},
+  {"zstd_compress_"        , (DL_FUNC) &zstd_compress_        , 4},
   {"zstd_decompress_"      , (DL_FUNC) &zstd_decompress_      , 1},
   {NULL, NULL, 0}
 };