+++ tests/testthat/test-cctx.R
@@ -1,0 +1,56 @@
+
+
+
+
+test_that("cctx with zstd_serialize works", {
+  
+  cctx <- init_zstd_cctx()
+  expect_identical(
+    zstd_serialize(mtcars),
+    zstd_serialize(mtcars, cctx = cctx)
+  )
+  
+  # Re-using a cctx
+  expect_identical(
+    zstd_serialize(mtcars),
+    zstd_serialize(mtcars, cctx = cctx)
+  )
+  
+})
+
+
+
+
+
+test_that("cctx with zstd_serialize_stream works", {
+  
+  cctx <- init_zstd_cctx()
+  expect_identical(
+    zstd_serialize_stream(mtcars),
+    zstd_serialize_stream(mtcars, cctx = cctx)
+  )
+  
+  # Re-using a cctx
+  expect_identical(
+    zstd_serialize_stream(mtcars),
+    zstd_serialize_stream(mtcars, cctx = cctx)
+  )
+  
+})
+
+
+test_that("cctx with zstd_compress works", {
+  
+  cctx <- init_zstd_cctx()
+  expect_identical(
+    zstd_compress(serialize(mtcars, NULL)),
+    zstd_compress(serialize(mtcars, NULL) , cctx = cctx)
+  )
+  
+  # Re-using a cctx
+  expect_identical(
+    zstd_compress(serialize(mtcars, NULL)),
+    zstd_compress(serialize(mtcars, NULL) , cctx = cctx)
+  )
+  
+})